requirements.txt
fastapi==0.131.0
uvicorn==0.41.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
